# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WALAB API Documentation",
        "description": "Documentation technique complète pour l'application Walab. Utilisez le bouton 'Authorize' pour tester les routes protégées.",
        "contact": {
            "email": "contact@walab.bj"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur API Principal"
        }
    ],
    "paths": {
...
        "/api/blog/list": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Liste des articles",
                "description": "Récupère tous les articles de blog, triés du plus récent au plus ancien.",
                "operationId": "23d1d9d93b25f7425a59ea23f0a7bea2",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des articles récupérée avec succès"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Blog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/blog/show/{id}": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Détail d'un article",
                "description": "Récupère un article spécifique avec ses images de galerie.",
                "operationId": "9d14c37f9fd01714b7dbdf5af44848dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Article récupéré"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Blog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article introuvable"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Blog": {
                "title": "Article de Blog",
                "description": "Modèle d'un article",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Les bienfaits du bilan sanguin"
                    },
                    "slug": {
                        "type": "string",
                        "example": "les-bienfaits-du-bilan-sanguin"
                    },
                    "meta_description": {
                        "type": "string",
                        "example": "Découvrez pourquoi il est important..."
                    },
                    "content": {
                        "type": "string",
                        "example": "<p>Contenu HTML...</p>"
                    },
                    "image_url": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/blogs/image.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez votre token d'accès ici (sans le préfixe 'Bearer')",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}